backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosqlite==0.19.0
greenlet>=3.0
anthropic==0.18.1
openai==1.12.0
